requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
curl_cffi==0.6.2
distlib==0.3.8
Django==5.0.4
django-environ==0.11.2
django-environment==0.3.8
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
filelock==3.13.4
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
packaging==24.0
pbr==6.0.0
platformdirs==4.2.0
pycparser==2.22
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
soupsieve==2.5
sqlparse==0.5.0
stevedore==5.2.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.0
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
yarl==1.9.4
